sales_hr/requirements.txt
langchain
langchain-community
langchain-openai
sentence-transformers
faiss-cpu
python-dotenv
pydantic
tiktoken
langchain_groq
langchain-huggingface